+++ apps/store/src/shared/composite/order-summary-aside.tsx
@@ -2,50 +2,7 @@ import { type LucideIcon } from "lucide-react";
 import { type ReactNode } from "react";
 import { cn } from "@/shared/lib/utils";
 import { UtilityKeyValueRow } from "@/shared/composite/utility-page";
-
-export interface SummaryRow {
-  id: string | number;
-  label: string;
-  value: ReactNode;
-  className?: string;
-}
-
-export interface PriceTotals {
-  originalPrice: number;
-  totalDiscount: number;
-  shippingCost: number;
-}
-
-export function buildPriceRows(totals: PriceTotals): SummaryRow[] {
-  return [
-    {
-      id: "original-price",
-      label: "상품 금액",
-      value: `${totals.originalPrice.toLocaleString()}원`,
-    },
-    ...(totals.totalDiscount > 0
-      ? [
-          {
-            id: "discount",
-            label: "할인 금액",
-            value: (
-              <span className="text-red-500">
-                -{totals.totalDiscount.toLocaleString()}원
-              </span>
-            ),
-          },
-        ]
-      : []),
-    {
-      id: "shipping",
-      label: "배송비",
-      value:
-        totals.shippingCost > 0
-          ? `${totals.shippingCost.toLocaleString()}원`
-          : "무료",
-    },
-  ];
-}
+import type { SummaryRow } from "./order-summary-utils";
 
 interface OrderSummaryAsideProps {
   title?: string;
